# Reduce overheads related to DictionaryBlock#getSizeInBytes()