--- a/core/trino-spi/src/main/java/io/trino/spi/block/AbstractRowBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/AbstractRowBlock.java
@@ -16,6 +16,7 @@ package io.trino.spi.block;
 import javax.annotation.Nullable;
 
 import java.util.List;
+import java.util.OptionalInt;
 
 import static io.trino.spi.block.BlockUtil.arraySame;
 import static io.trino.spi.block.BlockUtil.checkArrayRange;
@@ -131,6 +132,41 @@ public abstract class AbstractRowBlock
     }
 
     @Override
+    public final OptionalInt fixedSizeInBytesPerPosition()
+    {
+        if (!mayHaveNull()) {
+            // when null rows are present, we can't use the fixed field sizes to infer the correct
+            // size for arbitrary position selection
+            OptionalInt fieldSize = fixedSizeInBytesPerFieldPosition();
+            if (fieldSize.isPresent()) {
+                // must include the row block overhead in addition to the per position size in bytes
+                return OptionalInt.of(fieldSize.getAsInt() + (Integer.BYTES + Byte.BYTES)); // offsets + rowIsNull
+            }
+        }
+        return OptionalInt.empty();
+    }
+
+    /**
+     * Returns the combined {@link Block#fixedSizeInBytesPerPosition()} value for all fields, assuming all
+     * are fixed size. If any field is not fixed size, then no value will be returned. This does <i>not</i>
+     * include the size-per-position overhead associated with the {@link AbstractRowBlock} itself, only of
+     * the constituent field members.
+     */
+    private OptionalInt fixedSizeInBytesPerFieldPosition()
+    {
+        Block[] rawFieldBlocks = getRawFieldBlocks();
+        int fixedSizePerRow = 0;
+        for (int i = 0; i < numFields; i++) {
+            OptionalInt fieldFixedSize = rawFieldBlocks[i].fixedSizeInBytesPerPosition();
+            if (fieldFixedSize.isEmpty()) {
+                return OptionalInt.empty(); // found a block without a single per-position size
+            }
+            fixedSizePerRow += fieldFixedSize.getAsInt();
+        }
+        return OptionalInt.of(fixedSizePerRow);
+    }
+
+    @Override
     public long getRegionSizeInBytes(int position, int length)
     {
         int positionCount = getPositionCount();
@@ -148,27 +184,80 @@ public abstract class AbstractRowBlock
     }
 
     @Override
-    public long getPositionsSizeInBytes(boolean[] positions)
+    public final long getPositionsSizeInBytes(boolean[] positions, int selectedRowPositions)
     {
-        checkValidPositions(positions, getPositionCount());
+        int positionCount = getPositionCount();
+        checkValidPositions(positions, positionCount);
+        if (selectedRowPositions == 0) {
+            return 0;
+        }
+        if (selectedRowPositions == positionCount) {
+            return getSizeInBytes();
+        }
 
-        int usedPositionCount = 0;
-        boolean[] fieldPositions = new boolean[getRawFieldBlocks()[0].getPositionCount()];
-        for (int i = 0; i < positions.length; i++) {
-            if (positions[i]) {
-                usedPositionCount++;
-                int startFieldBlockOffset = getFieldBlockOffset(i);
-                int endFieldBlockOffset = getFieldBlockOffset(i + 1);
-                for (int j = startFieldBlockOffset; j < endFieldBlockOffset; j++) {
-                    fieldPositions[j] = true;
+        OptionalInt fixedSizePerFieldPosition = fixedSizeInBytesPerFieldPosition();
+        if (fixedSizePerFieldPosition.isPresent()) {
+            // All field blocks are fixed size per position, no specific position mapping is necessary
+            int selectedFieldPositionCount = selectedRowPositions;
+            boolean[] rowIsNull = getRowIsNull();
+            if (rowIsNull != null) {
+                // Some positions in usedPositions may be null which must be removed from the selectedFieldPositionCount
+                int offsetBase = getOffsetBase();
+                for (int i = 0; i < positions.length; i++) {
+                    if (positions[i] && rowIsNull[i + offsetBase]) {
+                        selectedFieldPositionCount--; // selected row is null, don't include it in the selected field positions
+                    }
+                }
+                if (selectedFieldPositionCount < 0) {
+                    throw new IllegalStateException("Invalid field position selection after nulls removed: " + selectedFieldPositionCount);
+                }
+            }
+            return ((Integer.BYTES + Byte.BYTES) * (long) selectedRowPositions) + (fixedSizePerFieldPosition.getAsInt() * (long) selectedFieldPositionCount);
+        }
+
+        // Fall back to specific position size calculations
+        return getSpecificPositionsSizeInBytes(positions, selectedRowPositions);
+    }
+
+    private long getSpecificPositionsSizeInBytes(boolean[] positions, int selectedRowPositions)
+    {
+        int positionCount = getPositionCount();
+        int offsetBase = getOffsetBase();
+        boolean[] rowIsNull = getRowIsNull();
+        // No fixed width size per row, specific positions used must be tracked
+        int totalFieldPositions = getRawFieldBlocks()[0].getPositionCount();
+        boolean[] fieldPositions;
+        int selectedFieldPositionCount;
+        if (rowIsNull == null) {
+            // No nulls, so the same number of positions are used
+            selectedFieldPositionCount = selectedRowPositions;
+            if (offsetBase == 0 && positionCount == totalFieldPositions) {
+                // No need to adapt the positions array at all, reuse it directly
+                fieldPositions = positions;
+            }
+            else {
+                // no nulls present, so we can just shift the positions array into alignment with the elements block with other positions unused
+                fieldPositions = new boolean[totalFieldPositions];
+                System.arraycopy(positions, 0, fieldPositions, offsetBase, positions.length);
+            }
+        }
+        else {
+            fieldPositions = new boolean[totalFieldPositions];
+            selectedFieldPositionCount = 0;
+            for (int i = 0; i < positions.length; i++) {
+                if (positions[i] && !rowIsNull[offsetBase + i]) {
+                    selectedFieldPositionCount++;
+                    fieldPositions[getFieldBlockOffset(i)] = true;
                 }
             }
         }
-        long sizeInBytes = 0;
+
+        Block[] rawFieldBlocks = getRawFieldBlocks();
+        long sizeInBytes = ((Integer.BYTES + Byte.BYTES) * (long) selectedRowPositions); // offsets + rowIsNull
         for (int j = 0; j < numFields; j++) {
-            sizeInBytes += getRawFieldBlocks()[j].getPositionsSizeInBytes(fieldPositions);
+            sizeInBytes += rawFieldBlocks[j].getPositionsSizeInBytes(fieldPositions, selectedFieldPositionCount);
         }
-        return sizeInBytes + (Integer.BYTES + Byte.BYTES) * (long) usedPositionCount;
+        return sizeInBytes;
     }
 
     @Override