--- a/core/trino-spi/src/main/java/io/trino/spi/block/AbstractSingleArrayBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/AbstractSingleArrayBlock.java
@@ -182,7 +182,7 @@ public abstract class AbstractSingleArrayBlock
     }
 
     @Override
-    public long getPositionsSizeInBytes(boolean[] positions)
+    public long getPositionsSizeInBytes(boolean[] positions, int selectedPositionsCount)
     {
         throw new UnsupportedOperationException();
     }