… core/trino-main/src/test/java/io/trino/block/AbstractTestBlock.java
… core/trino-main/src/test/java/io/trino/block/TestDictionaryBlock.java
… core/trino-main/src/test/java/io/trino/block/TestRunLengthEncodedBlock.java
… core/trino-main/src/test/java/io/trino/operator/project/BenchmarkDictionaryBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/Int128ArrayBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/Int128ArrayBlockBuilder.java
… core/trino-spi/src/main/java/io/trino/spi/block/Int96ArrayBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/Int96ArrayBlockBuilder.java
… core/trino-spi/src/main/java/io/trino/spi/block/IntArrayBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/IntArrayBlockBuilder.java
… core/trino-spi/src/main/java/io/trino/spi/block/LazyBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/LongArrayBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/LongArrayBlockBuilder.java
… core/trino-spi/src/main/java/io/trino/spi/block/RunLengthEncodedBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/ShortArrayBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/ShortArrayBlockBuilder.java
… core/trino-spi/src/main/java/io/trino/spi/block/SingleArrayBlockWriter.java
… core/trino-spi/src/main/java/io/trino/spi/block/SingleMapBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/SingleMapBlockWriter.java
… core/trino-spi/src/main/java/io/trino/spi/block/SingleRowBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/SingleRowBlockWriter.java
… core/trino-spi/src/main/java/io/trino/spi/block/VariableWidthBlock.java
… core/trino-spi/src/main/java/io/trino/spi/block/VariableWidthBlockBuilder.java