<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->

## Description
Reduces the overhead of calculating `DictionaryBlock#getSizeInBytes()` by making two changes to the `Block` interface definition:
- Adds a new argument `selectedPositionCount` to `Block#getPositionsSizeInBytes(boolean[] positions, int selectedPositionCount)`. Callers to the previous method `getPositionsSizeInBytes(boolean[] positions)` could trivially calculate and pass the total count of the positions selected, and failing to do so required the called blocks to count the number of `true` values in the positions array- sometimes repeatedly in the case of eg: `RowBlock`.
- Adds a new method `Block#fixedSizeInBytesPerPosition()` that allows blocks to describe whether their `getSizeInBytes()` can be calculated directly from the position count, ie: the size is not variable based on the specific positions selected

Also includes a change to eagerly populate the unique position count and size in bytes on the results of `DictionaryBlock#getPositions` when the result is not compact, since the selected positions array may have already been created  and would otherwise have to be reconstructed in a subsequent call to `DictionaryBlock#getSizeInBytes()`.

<!-- Elaborate beyond the title of the PR as necessary to help the reviewers and maintainers.-->

## General information

Is this change a fix, improvement, new feature, refactoring, or other?

_This is an improvement that reduces the overhead for common scenarios involving `DictionaryBlock#getSizeInBytes()`_

Is this a change to the core query engine, a connector, client library, or the SPI interfaces? (be specific)

_This change affects the SPI interface of blocks and refactors the implementations of `DictionaryBlock#getPositions` and `Block#getPositionsSizeInBytes` to leverage the new information available._

How would you describe this change to a non-technical end user or system administrator?

_Specifically describing this change to a non-technical audience should not be necessary_

## Related issues, pull requests, and links

<!-- List any issue that is fixed and provide links to other related PRs, upstream release notes, and other useful resources:
* Fixes #issuenumber
* Related documentation in #issuenumber
* [Some release notes](http://usefulinfo.example.com)
-->

<!-- The following sections are filled in by the maintainer with input from the contributor:
Use :white_check_mark: or (x) or whatever really to signal selection.
-->

## Documentation

( ) No documentation is needed.
( ) Sufficient documentation is included in this PR.
( ) Documentation PR is available with #prnumber.
( ) Documentation issue #issuenumber is filed, and can be handled later.

## Release notes

( ) No release notes entries required.
( ) Release notes entries required with the following suggested text:

```
# Section
* Fix some things. ({issue}`5678`)
```