@pettyjamesm 2022-02-08
Benchmark results
### Before
```
Benchmark                                                (selectedPositions)  (valueType)  Mode  Cnt     Score     Error  Units
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                  100      varchar  avgt    5   441.915 ±  39.146  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                  100      integer  avgt    5   357.674 ±  12.478  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                 1000      varchar  avgt    5   436.585 ±   5.135  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                 1000      integer  avgt    5   374.206 ±   4.870  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                10000      varchar  avgt    5   838.399 ±  17.185  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                10000      integer  avgt    5   647.638 ±   5.097  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes               100000      varchar  avgt    5  2437.982 ± 119.726  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes               100000      integer  avgt    5  1754.578 ±  87.259  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                         100      varchar  avgt    5   383.916 ±  9.507  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                         100      integer  avgt    5   329.267 ±  7.083  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                        1000      varchar  avgt    5   420.201 ±  4.594  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                        1000      integer  avgt    5   361.897 ± 28.052  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                       10000      varchar  avgt    5   836.371 ± 20.350  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                       10000      integer  avgt    5   636.434 ±  4.285  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                      100000      varchar  avgt    5  2489.630 ± 44.489  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                      100000      integer  avgt    5  1806.359 ± 24.145  us/op
```


### After
```
Benchmark                                                (selectedPositions)  (valueType)  Mode  Cnt     Score    Error  Units
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                  100      varchar  avgt    5   253.769 ± 10.734  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                  100      integer  avgt    5    56.284 ±  1.955  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                 1000      varchar  avgt    5   311.319 ±  7.131  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                 1000      integer  avgt    5    51.795 ±  0.427  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                10000      varchar  avgt    5   772.982 ± 13.421  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes                10000      integer  avgt    5   137.741 ±  8.286  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes               100000      varchar  avgt    5  2216.938 ± 45.850  us/op
BenchmarkDictionaryBlock.getPositionsThenGetSizeInBytes               100000      integer  avgt    5   623.966 ± 14.182  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                         100      varchar  avgt    5   303.344 ± 14.147  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                         100      integer  avgt    5    91.252 ± 42.979  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                        1000      varchar  avgt    5   374.223 ± 15.776  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                        1000      integer  avgt    5    85.253 ±  4.161  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                       10000      varchar  avgt    5   778.172 ± 16.958  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                       10000      integer  avgt    5   213.559 ±  1.584  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                      100000      varchar  avgt    5  2374.654 ± 10.486  us/op
BenchmarkDictionaryBlock.getPositionsSizeInBytes                      100000      integer  avgt    5   725.355 ± 26.155  us/op
```