@pettyjamesm 2022-02-24
More complete benchmark results were easier to produce in the Presto equivalent PR, and links to those results can be found in [my comment](https://github.com/prestodb/presto/pull/17349#issuecomment-1050190474) there.