@losipiuk 2022-02-25
@pettyjamesm The JMH benchmarks look nice! I have a question still. Did you have a chance to verify how much the change impacts end-to-end query execution? how much are we gaining in CPU/wall time over some verbose benchmark (e.g. tpc-h/ds), or over your production queries?
The change surely adds some complexity and it would be nice to have a good justification for adding that.