@pettyjamesm 2022-02-25
> Did you have a chance to verify how much the change impacts end-to-end query execution? how much are we gaining in CPU/wall time over some verbose benchmark (e.g. tpc-h/ds), or over your production queries?

I haven't had a chance to put these changes into our production environment, and in TPCDS / TPCH queries I wouldn't expect much of a measurable improvement at the macro level. The biggest gains are going to be in queries that make heavy use of `RowBlock` / `ArrayBlock` / `DictionaryBlock` combinations. I just tweaked the `BenchmarkUnnestOperator` implementation to call `operatorContext.addProcessedInput(inputPage.getSizeInBytes(), inputPage.getPositionCount())` as well as for the output page. Results in about ~45% throughput improvement on my laptop (eg: ~6.5 ops/second to ~10.0 ops/second)